refactor: extract translation service into services/translation.py (3618 → from 4103 line god file)
