+++ transaction_web_app.py
@@ -34,13 +34,32 @@ from typing import Optional
 import pandas as pd
 import streamlit as st
 
-GEMINI_PROVIDER = "gemini"
-OPENAI_PROVIDER = "openai"
-DEFAULT_GEMINI_MODEL = "gemini-3.1-flash-lite"
-DEFAULT_OPENAI_MODEL = "gpt-4o-mini"
-GEMINI_TRANSLATION_MODE = "AI-Powered (Gemini 3.1 Flash-Lite)"
-OPENAI_TRANSLATION_MODE = "AI-Powered (OpenAI)"
-LEGACY_OPENAI_TRANSLATION_MODE = "AI-Powered (GPT-3.5)"
+from services.translation import (
+    GEMINI_PROVIDER,
+    OPENAI_PROVIDER,
+    DEFAULT_GEMINI_MODEL,
+    DEFAULT_OPENAI_MODEL,
+    GEMINI_TRANSLATION_MODE,
+    OPENAI_TRANSLATION_MODE,
+    LEGACY_OPENAI_TRANSLATION_MODE,
+    _AI_USAGE_KEY,
+    _default_key_for,
+    _default_model_for,
+    _provider_from_mode,
+    _is_retryable_error,
+    _call_with_retry,
+    _record_usage,
+    _recover_partial_json,
+    _translate_batch_gemini_single_prompt,
+    normalize_japanese_text,
+    protect_known_merchants,
+    restore_known_merchants,
+    translate_japanese_to_english_fallback,
+    translate_japanese_to_english_ai,
+    translate_japanese_to_english_gemini,
+    translate_japanese_to_english,
+    translate_batch_ai,
+)
 
 # Data layer
 try:
@@ -457,510 +476,6 @@ def extract_transactions_from_image(
     return df
 
 
-def _streamlit_secret(section: str, key: str) -> Optional[str]:
-    """Read Streamlit secrets safely when running in or outside Streamlit."""
-    try:
-        value = st.secrets.get(section, {}).get(key)
-        return value or None
-    except Exception:
-        return None
-
-
-def _default_key_for(provider: str, explicit_key: str = None) -> Optional[str]:
-    """Resolve API keys from explicit input, Streamlit secrets, then env vars."""
-    if explicit_key:
-        return explicit_key
-
-    if provider == GEMINI_PROVIDER:
-        return (
-            _streamlit_secret("gemini", "api_key")
-            or os.getenv("GEMINI_API_KEY")
-            or os.getenv("GOOGLE_API_KEY")
-        )
-
-    return (
-        _streamlit_secret("openai", "api_key")
-        or os.getenv("OPENAI_API_KEY")
-    )
-
-
-def _default_model_for(provider: str, explicit_model: str = None) -> str:
-    if explicit_model:
-        return explicit_model
-    return DEFAULT_GEMINI_MODEL if provider == GEMINI_PROVIDER else DEFAULT_OPENAI_MODEL
-
-
-def _provider_from_mode(mode: str, model: str = None, base_url: str = None) -> str:
-    if base_url == GEMINI_PROVIDER or (model and model.startswith("gemini")):
-        return GEMINI_PROVIDER
-    if mode == GEMINI_TRANSLATION_MODE:
-        return GEMINI_PROVIDER
-    return OPENAI_PROVIDER
-
-
-def _is_retryable_error(exc: Exception) -> bool:
-    """Return True for transient API errors that are worth retrying."""
-    msg = str(exc).lower()
-    retryable_markers = (
-        "429",       # rate limit
-        "500",       # internal server error
-        "502",       # bad gateway
-        "503",       # service unavailable
-        "504",       # gateway timeout
-        "rate limit",
-        "quota",
-        "too many requests",
-        "timeout",
-        "connection",
-        "network",
-    )
-    return any(m in msg for m in retryable_markers)
-
-
-def _call_with_retry(fn, *args, max_attempts: int = 3, base_delay: float = 1.0, **kwargs):
-    """Call fn(*args, **kwargs) up to max_attempts times with exponential backoff.
-
-    Retries only on transient (_is_retryable_error) exceptions.  Re-raises on
-    the final attempt or on non-retryable errors.
-    """
-    import time
-    import random
-    last_exc = None
-    for attempt in range(max_attempts):
-        try:
-            return fn(*args, **kwargs)
-        except Exception as exc:
-            if not _is_retryable_error(exc):
-                raise
-            last_exc = exc
-            if attempt < max_attempts - 1:
-                delay = base_delay * (2 ** attempt) + random.uniform(0, 0.5)
-                time.sleep(delay)
-    raise last_exc
-
-
-_COST_PER_1M = {
-    # (input_usd, output_usd) per 1M tokens — free tier is $0 but we show paid rates for info
-    DEFAULT_GEMINI_MODEL: (0.25, 1.50),
-    DEFAULT_OPENAI_MODEL: (0.15, 0.60),
-}
-
-_AI_USAGE_KEY = "ai_usage"
-
-
-def _record_usage(
-    provider: str,
-    model: str,
-    input_tokens: int,
-    output_tokens: int,
-    was_fallback: bool = False,
-) -> None:
-    """Accumulate token/cost stats into st.session_state['ai_usage'] if available."""
-    try:
-        usage = st.session_state.setdefault(
-            _AI_USAGE_KEY,
-            {"calls": 0, "retries": 0, "fallbacks": 0, "input_tokens": 0, "output_tokens": 0, "cost_usd": 0.0, "provider": provider, "model": model},
-        )
-        usage["calls"] += 1
-        usage["input_tokens"] += input_tokens
-        usage["output_tokens"] += output_tokens
-        if was_fallback:
-            usage["fallbacks"] += 1
-        rates = _COST_PER_1M.get(model, (0.0, 0.0))
-        usage["cost_usd"] += (input_tokens * rates[0] + output_tokens * rates[1]) / 1_000_000
-        usage["provider"] = provider
-        usage["model"] = model
-    except Exception:
-        pass
-
-
-def translate_japanese_to_english_ai(text: str, api_key: str = None) -> str:
-    """Translate Japanese text to English using OpenAI for high accuracy."""
-    try:
-        # Normalize half-width to full-width etc. to improve translation quality
-        text_norm = normalize_japanese_text(text)
-        # Protect known merchant names with placeholders
-        protected_text, placeholders = protect_known_merchants(text_norm)
-        import openai
-        
-        # Check if text contains Japanese characters
-        if not protected_text or not any(ord(char) > 127 for char in protected_text):
-            return protected_text
-        
-        api_key = _default_key_for(OPENAI_PROVIDER, api_key)
-        
-        if not api_key:
-            st.warning("No OpenAI API key found. Using free translation fallback.")
-            return translate_japanese_to_english_fallback(text)
-        
-        # Configure OpenAI client
-        client = openai.OpenAI(api_key=api_key)
-        
-        # Create translation prompt
-        prompt = f"""
-        Translate the following Japanese text to English. This is from a credit card statement, so maintain accuracy for financial terms and merchant names.
-        
-        Japanese text: {protected_text}
-        
-        English translation:"""
-        
-        def _do_openai_call():
-            return client.chat.completions.create(
-                model=DEFAULT_OPENAI_MODEL,
-                messages=[
-                    {"role": "system", "content": "You are a professional translator specializing in financial documents. Translate Japanese to English accurately, especially for merchant names and financial terms."},
-                    {"role": "user", "content": prompt}
-                ],
-                max_tokens=100,
-                temperature=0.1,
-            )
-
-        response = _call_with_retry(_do_openai_call)
-        translated = response.choices[0].message.content.strip()
-        # Record token usage
-        try:
-            u = response.usage
-            _record_usage(OPENAI_PROVIDER, DEFAULT_OPENAI_MODEL, u.prompt_tokens, u.completion_tokens)
-        except Exception:
-            pass
-        # Restore protected merchant names
-        translated = restore_known_merchants(translated, placeholders)
-        return translated
-        
-    except Exception as e:
-        st.warning(f"AI translation failed for '{text}': {e}")
-        _record_usage(OPENAI_PROVIDER, DEFAULT_OPENAI_MODEL, 0, 0, was_fallback=True)
-        # Fallback to free translation
-        return translate_japanese_to_english_fallback(text)
-
-
-def translate_japanese_to_english_gemini(
-    text: str,
-    api_key: str = None,
-    model: str = None,
-) -> str:
-    """Translate Japanese text to English with Gemini Flash-Lite free-tier model."""
-    try:
-        text_norm = normalize_japanese_text(text)
-        protected_text, placeholders = protect_known_merchants(text_norm)
-
-        if not protected_text or not any(ord(char) > 127 for char in protected_text):
-            return protected_text
-
-        api_key = _default_key_for(GEMINI_PROVIDER, api_key)
-        if not api_key:
-            st.warning("No Gemini API key found. Using free translation fallback.")
-            return translate_japanese_to_english_fallback(text)
-
-        model = _default_model_for(GEMINI_PROVIDER, model)
-
-        from google import genai
-        from google.genai import types
-
-        config_kwargs = {
-            "temperature": 0.1,
-            "max_output_tokens": 120,
-        }
-        try:
-            config_kwargs["thinking_config"] = types.ThinkingConfig(thinking_budget=0)
-        except Exception:
-            pass
-
-        client = genai.Client(api_key=api_key)
-
-        def _do_gemini_call():
-            return client.models.generate_content(
-                model=model,
-                contents=(
-                    "Translate this Japanese credit-card statement merchant or memo "
-                    "to concise English. Preserve merchant names and financial terms. "
-                    "Return only the English translation.\n\n"
-                    f"Japanese text: {protected_text}"
-                ),
-                config=types.GenerateContentConfig(**config_kwargs),
-            )
-
-        response = _call_with_retry(_do_gemini_call)
-        translated = (getattr(response, "text", "") or "").strip()
-        if not translated:
-            raise RuntimeError("Gemini returned an empty translation")
-        # Record token usage
-        try:
-            um = response.usage_metadata
-            _record_usage(GEMINI_PROVIDER, model, um.prompt_token_count, um.candidates_token_count)
-        except Exception:
-            pass
-
-        return restore_known_merchants(translated, placeholders)
-    except Exception as e:
-        st.warning(f"Gemini translation failed for '{text}': {e}")
-        _record_usage(GEMINI_PROVIDER, model or DEFAULT_GEMINI_MODEL, 0, 0, was_fallback=True)
-        return translate_japanese_to_english_fallback(text)
-
-
-def _recover_partial_json(raw: str, index_to_text: dict) -> dict:
-    """Best-effort extraction of translations from a truncated/malformed JSON string.
-
-    The Gemini batch prompt asks for: {"0": "translation", "1": "translation", ...}
-    When JSON is valid we parse it directly.  When it is truncated we regex-extract
-    whatever "N": "value" pairs are present and fall back to the original text for
-    the rest.
-    """
-    import json
-
-    # Fast path: valid JSON
-    try:
-        parsed = json.loads(raw)
-        if isinstance(parsed, dict):
-            return {
-                index_to_text[str(k)]: str(v)
-                for k, v in parsed.items()
-                if str(k) in index_to_text
-            }
-    except (json.JSONDecodeError, ValueError):
-        pass
-
-    # Slow path: regex over truncated response
-    recovered = {}
-    for m in re.finditer(r'"(\d+)"\s*:\s*"((?:[^"\\]|\\.)*)"', raw):
-        idx, val = m.group(1), m.group(2)
-        if idx in index_to_text:
-            # Unescape basic JSON string escapes
-            try:
-                val = val.encode("raw_unicode_escape").decode("unicode_escape")
-            except Exception:
-                pass
-            recovered[index_to_text[idx]] = val
-
-    return recovered
-
-
-def _translate_batch_gemini_single_prompt(
-    texts: list,
-    api_key: str,
-    model: str,
-) -> dict:
-    """Send all texts in one Gemini prompt and return {original: translation}.
-
-    Falls back to empty dict on total failure so callers can retry per-item.
-    Recovery: if the JSON is truncated, regex-extracts whatever pairs are present
-    and returns originals for the rest (>=80% recovery on partial truncation).
-    """
-    if not texts:
-        return {}
-
-    try:
-        from google import genai
-        from google.genai import types
-
-        # Build an indexed map so we can match response indices back to originals
-        index_to_text = {str(i): t for i, t in enumerate(texts)}
-        numbered_lines = "\n".join(f'{i}: "{t}"' for i, t in enumerate(texts))
-
-        prompt = (
-            "Translate each numbered Japanese credit-card merchant / memo to English.\n"
-            "Return ONLY a JSON object: {\"0\": \"translation\", \"1\": \"translation\", ...}\n"
-            "Preserve proper nouns and brand names. No explanations.\n\n"
-            f"{numbered_lines}"
-        )
-
-        config_kwargs: dict = {
-            "temperature": 0.1,
-            "max_output_tokens": max(800, 80 * len(texts)),
-        }
-        try:
-            config_kwargs["thinking_config"] = types.ThinkingConfig(thinking_budget=0)
-        except Exception:
-            pass
-
-        client = genai.Client(api_key=api_key)
-
-        def _do_call():
-            return client.models.generate_content(
-                model=model,
-                contents=prompt,
-                config=types.GenerateContentConfig(**config_kwargs),
-            )
-
-        response = _call_with_retry(_do_call)
-        raw = (getattr(response, "text", "") or "").strip()
-        # Record token usage
-        try:
-            um = response.usage_metadata
-            _record_usage(GEMINI_PROVIDER, model, um.prompt_token_count, um.candidates_token_count)
-        except Exception:
-            pass
-        # Strip markdown code fences if present
-        raw = re.sub(r"^```(?:json)?\s*", "", raw, flags=re.MULTILINE)
-        raw = re.sub(r"\s*```$", "", raw, flags=re.MULTILINE)
-
-        result = _recover_partial_json(raw, index_to_text)
-        return result
-    except Exception:
-        return {}
-
-
-def translate_japanese_to_english_fallback(text: str) -> str:
-    """Fallback translation using deep-translator when AI translation fails."""
-    try:
-        # Normalize first to convert half-width Katakana to standard form
-        text_norm = normalize_japanese_text(text)
-        # Protect known merchant names
-        protected_text, placeholders = protect_known_merchants(text_norm)
-        from deep_translator import GoogleTranslator
-        if protected_text and any(ord(char) > 127 for char in protected_text):
-            translated = GoogleTranslator(source='ja', target='en').translate(protected_text)
-            translated = restore_known_merchants(translated, placeholders)
-            return translated
-        return protected_text
-    except Exception as e:
-        st.warning(f"Fallback translation failed for '{text}': {e}")
-        return text
-
-def translate_japanese_to_english(text: str, mode: str = "Free Fallback", api_key: str = None) -> str:
-    """Main translation function - handles different translation modes."""
-    if mode == GEMINI_TRANSLATION_MODE:
-        return translate_japanese_to_english_gemini(text, api_key)
-    if mode in (OPENAI_TRANSLATION_MODE, LEGACY_OPENAI_TRANSLATION_MODE):
-        return translate_japanese_to_english_ai(text, api_key)
-    elif mode == "Free Fallback":
-        return translate_japanese_to_english_fallback(text)
-    else:  # No Translation
-        return text
-
-
-def translate_batch_ai(
-    texts,
-    api_key: str = None,
-    model: str = None,
-    base_url: str = None,
-) -> dict:
-    """Translate a batch of descriptions through the selected AI provider.
-
-    1. Deduplicates the input list so each unique text is sent once.
-    2. Checks the persistent SQLite translation_cache first — cached texts cost zero API calls.
-    3. Sends only uncached texts to the provider.
-    4. Writes new translations back to the cache for future uploads.
-    """
-    provider = _provider_from_mode("", model=model, base_url=base_url)
-    mode = GEMINI_TRANSLATION_MODE if provider == GEMINI_PROVIDER else OPENAI_TRANSLATION_MODE
-    resolved_model = _default_model_for(provider, model)
-
-    # Unique, non-empty texts in input order
-    unique_texts = list(dict.fromkeys(str(t) for t in texts if str(t).strip()))
-
-    # 1. Check persistent DB cache
-    db_cached: dict = {}
-    if get_cached_translations is not None:
-        try:
-            db_cached = get_cached_translations(unique_texts)
-        except Exception:
-            pass
-
-    # 2. Only call provider for texts not in the DB cache
-    uncached = [t for t in unique_texts if t not in db_cached]
-    new_translations: dict = {}
-
-    if provider == GEMINI_PROVIDER and len(uncached) > 1:
-        # Send all uncached texts in one batch prompt with partial-JSON recovery
-        batch_result = _translate_batch_gemini_single_prompt(uncached, api_key, resolved_model)
-        new_translations.update(batch_result)
-        # Per-item fallback for any texts the batch prompt missed
-        for text in uncached:
-            if text not in new_translations:
-                new_translations[text] = translate_japanese_to_english_gemini(text, api_key, resolved_model)
-    else:
-        for text in uncached:
-            if provider == GEMINI_PROVIDER:
-                new_translations[text] = translate_japanese_to_english_gemini(text, api_key, resolved_model)
-            else:
-                new_translations[text] = translate_japanese_to_english(text, mode, api_key)
-
-    # 3. Persist new translations
-    if new_translations and save_translations is not None:
-        try:
-            save_translations(new_translations, model=resolved_model, provider=provider)
-        except Exception:
-            pass
-
-    # 4. Merge and return (DB hits + fresh translations)
-    return {**db_cached, **new_translations}
-
-
-def normalize_japanese_text(text: str) -> str:
-    try:
-        # Convert half-width kana to regular width, normalize compatibility characters
-        s = unicodedata.normalize('NFKC', text)
-        # Replace ASCII/Unicode hyphens between Katakana with prolonged sound mark 'ー'
-        s = _replace_hyphen_between_katakana(s)
-        # Collapse extra spaces
-        s = re.sub(r"\s+", " ", s).strip()
-        return s
-    except Exception:
-        return text
-
-
-def protect_known_merchants(text: str):
-    """Replace known JP merchant names with placeholders to avoid mistranslation.
-    Returns (processed_text, placeholders_dict).
-    """
-    merchant_map = {
-        'ドンキホーテ': 'Don Quijote',
-        'ドン・キホーテ': 'Don Quijote',
-        'ローソン': 'Lawson',
-        'セブンイレブン': '7-Eleven',
-        'ファミリーマート': 'FamilyMart',
-        'イオン': 'AEON',
-        'ニトリ': 'Nitori',
-        'マクドナルド': "McDonald's",
-        'ケンタッキー': 'KFC',
-        'スターバックス': 'Starbucks',
-        'イトーヨーカドー': 'Ito-Yokado',
-        '西友': 'Seiyu',
-        'ライフ': 'LIFE',
-    }
-    placeholders = {}
-    processed = text
-    idx = 0
-    for jp, en in merchant_map.items():
-        if jp in processed:
-            token = f"[[BRAND_{idx}]]"
-            processed = processed.replace(jp, token)
-            placeholders[token] = en
-            idx += 1
-    return processed, placeholders
-
-
-def restore_known_merchants(translated: str, placeholders: dict) -> str:
-    try:
-        restored = translated
-        for token, en in placeholders.items():
-            restored = restored.replace(token, en)
-        # Guard against specific bad expansion like "Don't" from Don-*
-        if "Don't" in restored and 'Don ' in restored.replace("Don't", 'Don '):
-            restored = restored.replace("Don't", 'Don')
-        return restored
-    except Exception:
-        return translated
-
-
-def _replace_hyphen_between_katakana(s: str) -> str:
-    """Replace hyphen-like chars between Katakana letters with the prolonged sound mark 'ー'.
-    Handles '-', '‐', '‑', '–', '—', and halfwidth 'ｰ'.
-    """
-    try:
-        # Katakana block \u30A0-\u30FF
-        hyphens = "-‐‑–—ｰ"
-        pattern = re.compile(rf"([\u30A0-\u30FF])[{hyphens}]([\u30A0-\u30FF])")
-        prev = None
-        out = s
-        # Iteratively replace until stable (for multiple hyphens)
-        while prev != out:
-            prev = out
-            out = pattern.sub(r"\1ー\2", out)
-        return out
-    except Exception:
-        return s
 
 def extract_transactions_from_csv(file_stream: io.BytesIO, translation_mode: str = "Free Fallback", api_key: str = None) -> pd.DataFrame:
     """Extract transactions from a CSV file.
